+++ lib/nav-bar-styles/neumorphic-bottom-nav-bar.widget.dart
@@ -19,6 +19,7 @@ class NeumorphicBottomNavBar extends StatelessWidget {
   final CurveType curveType;
   final NeumorphicProperties neumorphicProperties;
   final Function(int) popAllScreensForTheSelectedTab;
+  final bool popScreensOnTapOfSelectedTab;
 
   NeumorphicBottomNavBar(
       {Key key,
@@ -27,6 +28,7 @@ class NeumorphicBottomNavBar extends StatelessWidget {
       this.showElevation = false,
       this.iconSize,
       this.backgroundColor,
+      this.popScreensOnTapOfSelectedTab,
       this.animationDuration = const Duration(milliseconds: 1000),
       this.navBarHeight = 0.0,
       @required this.items,
@@ -160,7 +162,8 @@ class NeumorphicBottomNavBar extends StatelessWidget {
           return Flexible(
             child: GestureDetector(
               onTap: () {
-                if (this.previousIndex == index) {
+                if (this.popScreensOnTapOfSelectedTab &&
+                    this.previousIndex == index) {
                   this.popAllScreensForTheSelectedTab(index);
                 }
                 this.onItemSelected(index);