+++ lib/nav-bar-styles/style-7-bottom-nav-bar.widget.dart
@@ -17,6 +17,7 @@ class BottomNavStyle7 extends StatelessWidget {
   final double bottomPadding;
   final double horizontalPadding;
   final Function(int) popAllScreensForTheSelectedTab;
+  final bool popScreensOnTapOfSelectedTab;
 
   BottomNavStyle7(
       {Key key,
@@ -26,6 +27,7 @@ class BottomNavStyle7 extends StatelessWidget {
       this.iconSize,
       this.horizontalPadding,
       this.backgroundColor,
+      this.popScreensOnTapOfSelectedTab,
       this.animationDuration = const Duration(milliseconds: 270),
       this.navBarHeight = 0.0,
       @required this.items,
@@ -129,7 +131,8 @@ class BottomNavStyle7 extends StatelessWidget {
             child: GestureDetector(
               onTap: () {
                 this.onItemSelected(index);
-                if (this.previousIndex == index) {
+                if (this.popScreensOnTapOfSelectedTab &&
+                    this.previousIndex == index) {
                   this.popAllScreensForTheSelectedTab(index);
                 }
               },